refactor: the application is divided into modules
