+++ src/modules/index.ts
@@ -1,0 +1,4 @@
+export { moveCursorUp, moveCursorDown, moveCursorLeft, moveCursorRight } from './navigation'
+export { drawCircle, drawRectangle, drawSquare } from './drawing'
+export { takeScreenshot } from './printScreen'
+export { mainWrapper } from './common'
